Run MacOS Ventura In A VMware Workstation VM!
It took me a few days to get this working but I've gone ahead and written a Guide for how I did it for you all so I hope that it works for everyone that sees this at least! As it says in the ReadMe.md it works on my Hardware and my Hardware is pretty modern so I'm hoping it works on everyone else's!

I've only tested Ventura and am unsure if newer or older OS's work yet but I'd be willing to test and report back on that too but for now Ventura is the only one I can report a functioning in a VM with 128 MB of Display Graphics.